Fast Calculate Score Profile with Fine Time Scale

Calculates the multi-scale KL divergence score profile with fine time scales (0, 1, 2, 3, ...).

fast_calculate_score_profile_highres(
  W,
  max_time = 100,
  init_state,
  score_ls = c("score0")
)

Arguments

W

matrix; the affinity matrix.

max_time

integer; the maximum diffusion time. Default is 100. The actual maximum diffusion time may be shorter if all nodes converge beforehand.

init_state

matrix; the row-wise normalized initial state matrix.

score_ls

character vector; the list of scores to compute. See get_score_profile for details. Default is c("score0").

Value

A data frame containing the computed scores for different diffusion times.