Compute Score Profile

Computes various scores for state transitions based on different metrics such as KL divergence and entropy.

get_score_profile(
  state_0 = NULL,
  state_t = NULL,
  state_inf = NULL,
  P_diag_t = NULL,
  score_ls = c("score0", "delta_correction")
)

Arguments

state_0

matrix; the initial state matrix.

state_t

matrix; the state matrix at time t.

state_inf

matrix; the state matrix at equilibrium.

P_diag_t

matrix; the diagonal elements of the diffusion operator at time t.

score_ls

character vector; a list of scores to compute. Possible values are "score0", "max_score0", "score1", "delta_correction", and "entropy". Default is c("score0", "delta_correction").

"score0": KL divergence between the initial state (state_0) and the state at time t (state_t).

"max_score0": KL divergence between the initial state (state_0) and the equilibrium state (state_inf).

"score1": KL divergence between the state at time t (state_t) and the equilibrium state (state_inf).

"delta_correction": KL divergence between the initial state (state_0) and the diagonal elements of the diffusion operator at time t (diag(P^t)).

"entropy": Entropy of the state at time t (state_t).

Value

A data frame containing the computed scores, with each column representing a score.